+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Fixed
 ~~~~~
 
 * Allow tasks in the same transition with a "fail" command to run. (bug fix)
+* Fix Jinja block expression to render correctly. (bug fix)
 
 0.5
 ---